requirements.txt
streamlit>=1.36.0
streamlit-image-select>=0.6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.1
Pillow>=11.0.0
pytest>=8.0.0
google-cloud-storage>=2.16.0
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0
requests>=2.32.3
pandas>=2.2.0
google-api-python-client>=2.149.0
